Downgrade Serf to 0.7.0 in attempt to resolve cache issue
When making changes to APIs, these changes may not be reflected
until the cache is manually flushed. According to the community,
this is likely because of issues with Serf / the cluster.

Indeed, since Kong 0.10.1, Serf 0.7.0 is being used instead to
fix an issue with detecting the first non-loopback private IP address.

---

Also, remove unnecessary port expose, and print versions of
all core dependencies.